+++ src/bin/UserInput/Controls.h
@@ -30,6 +30,7 @@ public:
 		AddEmptyEntry,
 		MoveEntryForward,
 		MoveEntryBack,
+		PickUpSlot,
 		MoveWheelForward,
 		MoveWheelBack,
 		ToggleEditHints,